+++ salt/_modules/containerd.py
@@ -14,7 +14,7 @@ def __virtual__():
     return __virtualname__
 
 
-def load_cri_image(path):
+def load_cri_image(path, fullname=None):
     """
     Load a Docker image archive into the :program:`containerd` CRI image cache.
 
@@ -26,6 +26,7 @@ def load_cri_image(path):
         Path of the Docker image archive to load
     """
     log.info('Importing image from "%s" into CRI cache', path)
-    return __salt__["cmd.run_all"](
-        'ctr --debug -n k8s.io image import "{0}"'.format(path)
-    )
+    cmd = 'ctr --debug -n k8s.io image import "{}"'.format(path)
+    if fullname:
+        cmd += ' --index-name "{}"'.format(fullname)
+    return __salt__["cmd.run_all"](cmd)