+++ salt/_states/containerd.py
@@ -63,7 +63,9 @@ def image_managed(name, archive_path=None):
 
         real_archive_path = __salt__["cp.cache_file"](archive_path, __env__)
 
-        result = __salt__["containerd.load_cri_image"](path=real_archive_path)
+        result = __salt__["containerd.load_cri_image"](
+            path=real_archive_path, fullname=name
+        )
         # ctr can fail to load the image and exit silently
         if result["retcode"] == 0 and __salt__["cri.available"](name):
             ret["changes"].update(